Return proper DST information
The DST adjustment that is returned by the `FixedOffset' class is
currently the same as its UTC offset. This is not correct though, since
the DST adjustment should specify the daylight savings that are in
effect for a given datetime, usually 0 or 1 hour.

Since the `FixedOffset' class does not observe daylight savings time, it
should always return that the DST adjustment is 0. This is in accordance
with the definition of the `dst()' method of the Python documentation
[1].

Finally, update the unit tests to ensure that the proper UTC offset and
DST adjustment are always returned.

[1] https://docs.python.org/2/library/datetime.html#datetime.tzinfo.dst